+++ aspgen/__init__.py
@@ -1,4 +1,5 @@
 #! /usr/bin/env python
+# -*- coding: UTF-8 -*-
 
 """Initialize aspgen package
 
@@ -28,4 +29,4 @@ __email__ = 'theonehyer@gmail.com'
 __license__ = 'GPLv3'
 __maintainer__ = 'Alex Hyer'
 __status__ = 'Alpha'
-__version__ = '1.2.0a4'
+__version__ = '1.2.0a5'